--- a/export.py
+++ b/export.py
@@ -145,6 +145,7 @@ def export_saved_model(model, im, file, dynamic,
         inputs = keras.Input(shape=(*imgsz, 3), batch_size=None if dynamic else batch_size)
         outputs = tf_model.predict(inputs, tf_nms, agnostic_nms, topk_per_class, topk_all, iou_thres, conf_thres)
         keras_model = keras.Model(inputs=inputs, outputs=outputs)
+        keras_model.trainable = False
         keras_model.summary()
         keras_model.save(f, save_format='tf')