@glenn-jocher 2021-09-16
@zldrobit can confirm super speedup on TFLite detect.py inference:

## Before
```python
detect: weights=['yolov5s.tflite'], source=data/images, imgsz=[640, 640], conf_thres=0.25, iou_thres=0.45, max_det=1000, device=, view_img=False, save_txt=False, save_conf=False, save_crop=False, nosave=False, classes=None, agnostic_nms=False, augment=False, visualize=False, update=False, project=runs/detect, name=exp, exist_ok=False, line_thickness=3, hide_labels=False, hide_conf=False, half=False
YOLOv5 🚀 v5.0-436-g6b44ecd torch 1.9.0+cu102 CUDA:0 (Tesla T4, 15109.75MB)

image 1/2 /content/yolov5/data/images/bus.jpg: 640x640 4 class0s, 1 class5, Done. (23.460s)
image 2/2 /content/yolov5/data/images/zidane.jpg: 640x640 2 class0s, 2 class27s, Done. (23.573s)
Speed: 5.0ms pre-process, 23516.5ms inference, 8.4ms NMS per image at shape (1, 3, 640, 640)
Results saved to runs/detect/exp
```

## This PR
```python
detect: weights=['yolov5s-fp16.tflite'], source=data/images, imgsz=[640, 640], conf_thres=0.25, iou_thres=0.45, max_det=1000, device=, view_img=False, save_txt=False, save_conf=False, save_crop=False, nosave=False, classes=None, agnostic_nms=False, augment=False, visualize=False, update=False, project=runs/detect, name=exp, exist_ok=False, line_thickness=3, hide_labels=False, hide_conf=False, half=False
YOLOv5 🚀 v3.0-901-gfe2b1ec torch 1.9.0+cu102 CUDA:0 (Tesla T4, 15109.75MB)

image 1/2 /content/yolov5/data/images/bus.jpg: 640x640 4 class0s, 1 class5, Done. (0.403s)
image 2/2 /content/yolov5/data/images/zidane.jpg: 640x640 2 class0s, 2 class27s, Done. (0.320s)
Speed: 4.8ms pre-process, 361.5ms inference, 7.7ms NMS per image at shape (1, 3, 640, 640)
Results saved to runs/detect/exp
```