@glenn-jocher 2021-09-16
@zldrobit also confirm trainable params are now 0:

```
Total params: 7,266,973
Trainable params: 0
Non-trainable params: 7,266,973
```