@alexdwu13 2021-09-24
@zldrobit now that the Tflite export defaults to FP16 post-quantization, how does the `--half` argument fit in? Is there a difference to enabling the flag vs. leaving it off?