@glenn-jocher 2021-09-24
@alexdwu13 that's a good question. `--half` is really meant for exporting other formats like ONNX in FP16. I'm not sure what effect, if any, `--half` has on TF/TFLite/TFjs models. Have you exported with and without --half to compare?