@alexdwu13 2021-09-25
@glenn-jocher so I've actually been unable to successfully use the `--half` option as it requires `--device 0`, which breaks for me both on Collab and in my local environment. Seems like this bug has been addressed – though the fix does not seem applicable to the current code https://github.com/ultralytics/yolov5/issues/2106 


But I did compare `tf.float16` vs. default (`tf.float32`) by commenting line 191 out in `yolov5/export.py`:
```
# converter.target_spec.supported_types = [tf.float16]
```
If you drop the 2 models into https://netron.app/ you can see that `yolov5s6-fp16.tflite` has a bunch of **Dequantize** layers compared to `yolov5s6-fp32.tflite`. Additionally, the FP16 model runs ~20% slower on TFLite GPU Delegate due to the additional **Dequantize** layers running on CPU:

```
$ ./android_aarch64_benchmark_model --use_gpu=true --graph=yolov5s6-fp32.tflite                                                                                    
STARTING!
Log parameter values verbosely: [0]
Graph: [yolov5s6-fp32.tflite]
Use gpu: [1]
Loaded model yolov5s6-fp32.tflite
INFO: Initialized TensorFlow Lite runtime.
INFO: Created TensorFlow Lite delegate for GPU.
INFO: Initialized OpenCL-based API.
INFO: Created 1 GPU delegate kernels.
Explicitly applied GPU delegate, and the model graph will be completely executed by the delegate.
The input model file size (MB): 13.2676
Initialized session in 2974.27ms.
Running benchmark for at least 1 iterations and at least 0.5 seconds but terminate if exceeding 150 seconds.
count=3 first=219376 curr=202144 min=197099 max=219376 avg=206206 std=9537

Running benchmark for at least 50 iterations and at least 1 seconds but terminate if exceeding 150 seconds.
count=50 first=202314 curr=198963 min=197728 max=204225 avg=201360 std=1733
```

```
$ ./android_aarch64_benchmark_model --use_gpu=true --graph=yolov5s6-fp16.tflite                                                                                    
STARTING!
Log parameter values verbosely: [0]
Graph: [yolov5s6-fp16.tflite]
Use gpu: [1]
Loaded model yolov5s6-fp16.tflite
INFO: Initialized TensorFlow Lite runtime.
INFO: Created TensorFlow Lite delegate for GPU.
ERROR: Following operations are not supported by GPU delegate:
DEQUANTIZE: 
521 operations will run on the GPU, and the remaining 5 operations will run on the CPU.
INFO: Initialized OpenCL-based API.
INFO: Created 1 GPU delegate kernels.
Explicitly applied GPU delegate, and the model graph will be partially executed by the delegate w/ 1 delegate kernels.
The input model file size (MB): 25.5305
Initialized session in 1865.52ms.
Running benchmark for at least 1 iterations and at least 0.5 seconds but terminate if exceeding 150 seconds.
count=3 first=228432 curr=253663 min=228432 max=253663 avg=239246 std=10610

Running benchmark for at least 50 iterations and at least 1 seconds but terminate if exceeding 150 seconds.
count=50 first=252072 curr=248613 min=235499 max=256871 avg=251860 std=4912
```

Since the GPU delegate should be able to natively run in FP16 it seems strange these `dequantize` ops are added. I wonder if  they'd go away if the model was exported with `--half --device 0`...

[yolov5s6-fp32.tflite.zip](https://github.com/ultralytics/yolov5/files/7228891/yolov5s6-fp32.tflite.zip)
[yolov5s6-fp16.tflite.zip](https://github.com/ultralytics/yolov5/files/7228892/yolov5s6-fp16.tflite.zip)