@glenn-jocher 2021-09-25
@alexdwu13 yes it's true that there is an assert to avoid using --half with cpu. This is because pytorch is unable to run CPU inference with FP16 models, and we do dry runs with the pytorch model here to build grids for example:
https://github.com/ultralytics/yolov5/blob/2c2ef25f8bb351b34aef89f8fce75742c698e847/export.py#L296-L298

Maybe we could cast the model and image to half after this pytorch inference?