@glenn-jocher 2021-09-25
@alexdwu13 ok I ran an experiment. If I cast to .half() after pytorch inference I get errors on TorchScript and ONNX export:

```python
PyTorch: starting from /Users/glennjocher/PycharmProjects/yolov5/yolov5s.pt (14.8 MB)

TorchScript: starting export with torch 1.9.1...
TorchScript: export failure: "unfolded2d_copy" not implemented for 'Half'

ONNX: starting export with onnx 1.10.1...
ONNX: export failure: "unfolded2d_copy" not implemented for 'Half'
```

But TFLite export works fine, though the export -fp16.tflite model still has dequantize blocks in it. All of the actual Conv layers are defined in FP32 interestingly (in the -fp16.tflite model), so yes it seems like providing the model directly in FP32 is most efficient unless there is a way to natively have the Conv layers exist natively in FP16. @zldrobit what do you think?

<img width="1284" alt="Screen Shot 2021-09-24 at 7 22 41 PM" src="https://user-images.githubusercontent.com/26833433/134754713-7d5d5c7c-461e-4eab-ba93-f7fe7e8b6f56.png">